services/requirements.txt
psycopg2-binary
sentry-sdk
mwclient
mwparserfromhell
unidecode
langid
gensim
requests
boto3
pySqsListener